Process: Fix printf modifier warning
Another cast, for the %*s which expects a signed value to our unsigned
length parameter.

In this instance, it feels like the printf modifier type is wrong,
(surely you can't have a negative field length here?) but I'm not
one to argue with the C-Library

+++ src/process.c
@@ -4863,7 +4863,7 @@ static void ColonFin(char *buf, size_t len, void *data)
 			}
 		}
 		if (l > r && showmessage)
-			Msg(0, "No commands matching '%*s'", len, buf);
+			Msg(0, "No commands matching '%*s'", (int)len, buf);
 		return;
 	}
